feat(webclient): full-page /feedback/new entry point with shareable URL state
Extract the add-entry tab UI out of AddProposalModal into a reusable
AddProposalForm and mount it on a new /feedback/new page. The page seeds
the draft from `?kind=` and `?org=` query parameters and mirrors edits back
to the URL so links like `/feedback/new?kind=event&org=45411` open the form
with the event tab and "Studentische Initiativen" pre-selected.

Make useKnownOrgs expose an awaitable `ready()` so the page can block setup
on the orgs CDN only when a pre-set org needs to resolve on first render -
other entry points keep the existing lazy fetch.
